pyproject.toml
[tool.poetry]
name = "document-ai-agent"
version = "0.1.0"
description = "Локальный AI агент на базе Ollama для работы с нормативной документацией по закупкам"
authors = ["AI Assistant <ai@example.com>"]
readme = "README.md"
packages = [{include = "ai_agent"}]

[tool.poetry.dependencies]
python = "^3.9"
ollama = "^0.3.0"
chromadb = "^0.4.0"
click = "^8.1.0"
rich = "^13.0.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
python-docx = "^1.1.0"
numpy = "<2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
docai = "ai_agent.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"